moved processing rows into data from pkg excel to pkg jobdata
